backend/requirements.txt
amqp==2.6.0
asgiref==3.2.7
billiard==3.6.3.0
celery==4.4.4
Django==3.0.6
django-cors-headers==3.3.0
django-filter==2.2.0
django-rest-passwordreset==1.1.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
importlib-metadata==1.6.0
kombu==4.6.10
PyJWT==1.7.1
pytz==2020.1
redis==3.5.3
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
vine==1.3.0
zipp==3.1.0
